chore(runtime): split object prototype helpers
